Refactoring, changes for resume support
